Fix bug in OtbReader::NextU32
The function used an ushort cast instead of a uint cast

+++ Backend/OtbReader.cs
@@ -68,7 +68,7 @@ namespace Backend.Otb
                     ++cursor;
                 }
 
-                result += (ushort)(((uint)buffer[cursor]) << (8 * shift));
+                result += (uint)(((uint)buffer[cursor]) << (8 * shift));
                 ++cursor;
                 ++shift;
             }
